fix: solve multiple requests and improve code splitting
