+++ src/containers/Invoice/Detail/Detail.jsx
@@ -12,24 +12,26 @@ import Stats from "./Stats";
 import get from "lodash/fp/get";
 import { useDocumentTitle } from "src/hooks/utils/useDocumentTitle";
 import { presentationalInvoiceName } from "../helpers";
+import { isUserDeveloper } from "src/containers/DCC/helpers";
 
 const InvoiceDetail = ({ Main, match }) => {
   const invoiceToken = get("params.token", match);
   const threadParentCommentID = get("params.commentid", match);
+
   const {
     invoice,
     loading,
     currentUser,
     error,
     proposals,
+    user,
     proposalsError
   } = useInvoice(invoiceToken);
+
   const isAuthor =
     currentUser && invoice && invoice.userid === currentUser.userid;
   const isAdmin = currentUser && currentUser.isadmin;
   const isPublicMode = !isAdmin && !isAuthor;
-
-  const shouldShowStats = isAdmin && invoice;
 
   // set tab title
   useDocumentTitle(presentationalInvoiceName(invoice));
@@ -42,18 +44,26 @@ const InvoiceDetail = ({ Main, match }) => {
           {error ? (
             <Message kind="error">{error.toString()}</Message>
           ) : !!invoice && !loading ? (
-            <Invoice
-              invoice={invoice}
-              isAuthor={isAuthor}
-              isPublicMode={isPublicMode}
-              extended
-              approvedProposals={proposals || []}
-              approvedProposalsError={proposalsError}
-            />
+            <>
+              <Invoice
+                invoice={invoice}
+                isAuthor={isAuthor}
+                isPublicMode={isPublicMode}
+                extended
+                approvedProposals={proposals || []}
+                approvedProposalsError={proposalsError}
+              />
+              {isAdmin && (
+                <Stats
+                  invoiceToken={invoice.censorshiprecord.token}
+                  userid={invoice.userid}
+                  isUserDeveloper={isUserDeveloper(user)}
+                />
+              )}
+            </>
           ) : (
             <InvoiceLoader extended />
           )}
-          {shouldShowStats && <Stats invoice={invoice} />}
           {!isPublicMode && (
             <Comments
               recordAuthorID={invoice && invoice.userid}
