+++ src/containers/Invoice/Detail/hooks.js
@@ -3,8 +3,6 @@ import * as sel from "src/selectors";
 import * as act from "src/actions";
 import { useSelector, useAction } from "src/redux";
 import useAPIAction from "src/hooks/utils/useAPIAction";
-import useUserDetail from "src/hooks/api/useUserDetail";
-import { isUserDeveloper } from "src/containers/DCC/helpers";
 import useFetchMachine from "src/hooks/utils/useFetchMachine";
 import { getProposalsTokensFromInvoice } from "../helpers";
 import isEmpty from "lodash/fp/isEmpty";
@@ -29,6 +27,12 @@ export function useInvoice(invoiceToken) {
   )(keys(proposalsByToken));
 
   const hasUnfetchedProposalsTokens = !isEmpty(unfetchedProposalsTokens);
+
+  const userSelector = useMemo(
+    () => sel.makeGetUserByID(invoice && invoice.userid),
+    [invoice]
+  );
+  const user = useSelector(userSelector);
 
   const initialValues = {
     status: "idle",
@@ -79,7 +83,8 @@ export function useInvoice(invoiceToken) {
     loading: state.loading || state.verifying,
     proposals: state.proposals,
     proposalsError: state.error,
-    currentUser
+    currentUser,
+    user
   };
 }
 
@@ -87,8 +92,6 @@ export function useInvoicesSummary(currentInvoice, userid, start, end) {
   const onFetchAdminInvoicesWithoutState = useAction(
     act.onFetchAdminInvoicesWithoutState
   );
-
-  const { user } = useUserDetail(userid);
 
   const [
     loading,
@@ -99,8 +102,6 @@ export function useInvoicesSummary(currentInvoice, userid, start, end) {
   return {
     loading,
     error,
-    // return if the user is developer
-    isUserDeveloper: isUserDeveloper(user),
     // return invoices array but filter the current invoice out
     invoices: invoices
       ? invoices.filter((inv) => inv.censorshiprecord.token !== currentInvoice)
